rename: rename folder to sendEmailControllers
